hdl fpga: add SPI master peripheral
Port 0x1F is attached to the control register, which handle SPI chip
select outputs.  Bits 1..0 of this register can be:

	0 - nothing selected
	1 - on-board SPI flash (CSO)
	2 - user connector TVS0
	3 - user connector TVS1

Port 0x3F is attached to the data register.  Writing a value starts an
SPI transaction.  Reading returns the current contents of the shift
register (i.e., last read SPI byte if the controller is idle).